Add query spec file with initializer and getters
I want to distinguish between stateful elements in the queries and
stateless. Stateless elements are the same in every sub-query; stateful
elements change depending on which sub-query is being performed.

The "query spec" contains all information about the global query. It's
essentially a serialization of the arguments to `aw_freeform_table`,
with certain guarantees about the contents. For example, `limit`,
`page`, `dimensions`, and `sort` are always all the same length.

With a set of getter functions, you don't need to know exactly how the
data structure is built. Also, this is more resilient to changes over
time. It's a little clunky, but it's also safe.

The getter functions are all prefixed with `qs` for query spec.

This is the first step towards Issue #122.