Deprecate segmentRsid argument in aw_segment_table
This argument limited the usefulness of the function, without adding
anything in return. Previously I thought that `aw_get_segments` required
a report suite to work, and so I used to use that argument. Using NULL
is better, and causes all report suites to be searched. Since segment
IDs are unique, there is no loss of accuracy in searching all report
suites.

Resolves an issue raised by #149.